게임 등록과 수정에서 공통으로 사용되는 Form 모듈화하여 중복 제거
