+++ pages/game/add.tsx
@@ -1,26 +1,15 @@
-import { ChangeEvent, useState } from "react";
+import { useState } from "react";
 import { useRouter } from "next/router";
 import Head from "next/head";
-import {
-  Button,
-  Dimmer,
-  Form,
-  Header,
-  Icon,
-  List,
-  Loader,
-  Message,
-} from "semantic-ui-react";
+import { Dimmer, Header, Loader } from "semantic-ui-react";
 
 import styles from "../../styles/game/add.module.css";
 
 import { useUserContext } from "../../src/common/contexts/UserContext";
-import {
-  checkGid,
-  checkPid,
-  checkYoutubeUrl,
-} from "../../src/common/utils/regex";
-import { uploadImageToFirebaseStorage } from "../../src/common/utils/file";
+import useGameForm from "../../src/common/hooks/useGameForm";
+import { replaceLineBreakWithBrTag } from "../../src/common/utils/game";
+
+import GameForm from "../../src/common/components/GameForm";
 
 const GameAddPage = () => {
   const router = useRouter();
@@ -28,57 +17,11 @@ const GameAddPage = () => {
     state: { user },
   } = useUserContext();
 
-  const [title, setTitle] = useState("");
-  const [gid, setGid] = useState("");
-  const [pid, setPid] = useState("");
-  const [content, setContent] = useState("");
-  const [tags, setTags] = useState("");
-  const [uploadingImage, setUploadingImage] = useState(false);
-  const [images, setImages] = useState([]);
-  const [youtubeUrl, setYoutubeUrl] = useState("");
-  const [maker, setMaker] = useState("");
-  const [source, setSource] = useState("");
+  const gameFormState = useGameForm();
+  const { title, gid, pid, content, tags, images, youtubeUrl, maker, source } =
+    gameFormState;
+
   const [isSubmit, setIsSubmit] = useState(false);
-
-  const uploadImages = async (e: ChangeEvent<HTMLInputElement>) => {
-    const { files } = e.target;
-    if (files.length > 0) {
-      setUploadingImage(true);
-      try {
-        const promises = [...files].map(async (file) => {
-          const url = await uploadImageToFirebaseStorage(user.uid, file);
-          setImages((images) => [...images, { originName: file.name, url }]);
-        });
-        await Promise.all(promises);
-      } catch (error) {
-        alert("이미지 업로드 도중 에러가 발생했어요!");
-      } finally {
-        setUploadingImage(false);
-      }
-    }
-  };
-
-  const moveImageUp = (index: number) => {
-    setImages((images) => {
-      const newImage = [...images];
-      [newImage[index], newImage[index - 1]] = [
-        newImage[index - 1],
-        newImage[index],
-      ];
-      return newImage;
-    });
-  };
-
-  const moveImageDown = (index: number) => {
-    setImages((images) => {
-      const newImage = [...images];
-      [newImage[index], newImage[index + 1]] = [
-        newImage[index + 1],
-        newImage[index],
-      ];
-      return newImage;
-    });
-  };
 
   const submit = async () => {
     if (!title || !gid || isSubmit) return;
@@ -95,7 +38,7 @@ const GameAddPage = () => {
           title,
           gid,
           pid,
-          content: content.replace(/\r\n|\r|\n/g, "<br />"),
+          content: replaceLineBreakWithBrTag(content),
           tags: tags.trim() ? tags.trim().split(" ") : [],
           images,
           youtubeUrl,
@@ -120,10 +63,6 @@ const GameAddPage = () => {
     }
   };
 
-  const removeImage = (index: number) => {
-    setImages((images) => images.filter((_, i) => index !== i));
-  };
-
   if (!user)
     return (
       <Dimmer active>
@@ -138,168 +77,11 @@ const GameAddPage = () => {
         <meta property="og:title" content="게임 등록 - 차차코 게임 공유" />
       </Head>
       <Header size="huge">게임 등록</Header>
-      <Form onSubmit={submit}>
-        <Form.Field>
-          <label>
-            게임 제목 <span className={styles.gray}>(필수)</span>
-          </label>
-          <input
-            placeholder="게임 제목을 입력해주세요"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            게임 ID <span className={styles.gray}>(필수)</span>
-          </label>
-          <Form.Input
-            fluid
-            error={
-              gid && !checkGid(gid) ? "게임 ID 형식이 올바르지 않습니다." : null
-            }
-            placeholder="G-000-000-000"
-            value={gid}
-            onChange={(e) => setGid(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            프로그래머 ID <span className={styles.gray}>(선택)</span>
-          </label>
-          <Form.Input
-            fluid
-            error={
-              pid && !checkPid(pid)
-                ? "프로그래머 ID 형식이 올바르지 않습니다."
-                : null
-            }
-            placeholder="P-000-000-000"
-            value={pid}
-            onChange={(e) => setPid(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            게임 설명 <span className={styles.gray}>(선택)</span>
-          </label>
-          <textarea
-            placeholder="게임 설명을 입력해주세요"
-            value={content}
-            onChange={(e) => setContent(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            태그 입력{" "}
-            <span className={styles.gray}>(선택) (띄어쓰기로 구분)</span>
-          </label>
-          <Form.Input
-            fluid
-            placeholder="2D 3D 1인칭_FPS 격투 등"
-            value={tags}
-            onChange={(e) => setTags(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            이미지 추가{" "}
-            <span className={styles.gray}>
-              (선택) (16:9 사이즈 권장) (첫번째 이미지를 썸네일로 이용)
-            </span>
-          </label>
-          <input
-            type="file"
-            name="file"
-            onChange={uploadImages}
-            accept="image/*"
-            multiple
-          />
-        </Form.Field>
-        <List as="ol">
-          {images.map((image, index) => (
-            <List.Item key={image.originName} as="li">
-              <List.Content verticalAlign="middle">
-                {index === 0 ? "(썸네일) " : ""}
-                {image.originName}
-                {index > 0 && (
-                  <Icon
-                    name="arrow alternate circle up outline"
-                    className={styles.arrowButton}
-                    onClick={() => moveImageUp(index)}
-                  />
-                )}
-                {images.length > 1 && index < images.length - 1 && (
-                  <Icon
-                    name="arrow alternate circle down outline"
-                    className={styles.arrowButton}
-                    onClick={() => moveImageDown(index)}
-                  />
-                )}
-                <Icon
-                  name="remove circle"
-                  className={styles.removeIcon}
-                  onClick={() => removeImage(index)}
-                />
-              </List.Content>
-            </List.Item>
-          ))}
-        </List>
-        {uploadingImage && <Message>이미지를 업로드하는 중입니다...</Message>}
-        <Form.Field>
-          <label>
-            유튜브 영상 <span className={styles.gray}>(선택)</span>
-          </label>
-          <Form.Input
-            fluid
-            error={
-              youtubeUrl && !checkYoutubeUrl(youtubeUrl)
-                ? "유튜브 URL 형식이 올바르지 않습니다."
-                : null
-            }
-            placeholder="유튜브 영상의 링크를 입력해주세요"
-            value={youtubeUrl}
-            onChange={(e) => setYoutubeUrl(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            제작자 <span className={styles.gray}>(선택)</span>
-          </label>
-          <Form.Input
-            fluid
-            placeholder="제작자의 이름을 입력해주세요"
-            value={maker}
-            onChange={(e) => setMaker(e.target.value)}
-          />
-        </Form.Field>
-        <Form.Field>
-          <label>
-            출처 <span className={styles.gray}>(선택)</span>
-          </label>
-          <Form.Input
-            fluid
-            placeholder="출처를 입력해주세요"
-            value={source}
-            onChange={(e) => setSource(e.target.value)}
-          />
-        </Form.Field>
-        <Button
-          type="submit"
-          color="yellow"
-          disabled={
-            !title ||
-            !gid ||
-            uploadingImage ||
-            !checkGid(gid) ||
-            (pid && !checkPid(pid)) ||
-            (youtubeUrl && !checkYoutubeUrl(youtubeUrl))
-          }
-          loading={isSubmit}
-        >
-          등록
-        </Button>
-      </Form>
+      <GameForm
+        gameFormState={gameFormState}
+        submit={submit}
+        isSubmit={isSubmit}
+      />
     </div>
   );
 };
