+++ styles/game/add.module.css
@@ -7,25 +7,6 @@
   justify-content: center;
 }
 
-.fileIcon {
-  padding: 0 !important;
-}
-
-.removeIcon {
-  padding-left: 1em;
-  cursor: pointer;
-}
-
-.arrowButton {
-  margin-left: 0.5em !important;
-  cursor: pointer;
-}
-
-span.gray {
-  color: rgba(0, 0, 0, 0.5) !important;
-  font-weight: normal;
-}
-
 @media screen and (max-width: 700px) {
   .container {
     height: auto;
